Configurable allocator
