+++ Cargo.toml
@@ -14,7 +14,7 @@ license = "MIT"
 [features]
 mock_shutdown = []
 metered_allocator = [ "prometheus" ]
-mimalloc = [ "deps:mimalloc" ]
+mimalloc = [ "dep:mimalloc" ]
 rand = [ "dep:rand", "dep:rand_chacha" ]
 rayon = [ "dep:rayon", "dep:num_cpus" ]
 prometheus = [ "dep:prometheus", "dep:hyper", "dep:url" ]
@@ -29,7 +29,6 @@ futures = "0.3"
 hex = "0.4.3"
 hex-literal = "0.3"
 itertools = "0.10"
-mimalloc = { version = "0.1", default-features = false, optional = true }
 once_cell = "1.8"
 proptest = { version = "1.0", optional = true }
 structopt = "0.3"
@@ -39,6 +38,9 @@ tracing = "0.1"
 tracing-futures = "0.2"
 tracing-subscriber = { version = "0.3", features = [ "env-filter", "json" ] }
 users = "0.11"
+
+# Mimalloc feature
+mimalloc = { version = "0.1", optional = true }
 
 # Rand feature
 rand = { version = "0.8.5", optional = true }
